32-bit DAWs and Windows on ARM
- One architecture per build tree (cmake -A x64 | Win32 | ARM64); ASIO
  DLLs named ...ASIO64 / ...ASIO32 / ...ASIOARM64, same CLSIDs.
- 32-bit: registered in the 32-bit registry view. Shared memory and the
  Virtual Cable IOCTL do not depend on pointer size; an ASIO size check
  and _BitScanForward64 did.
- Windows on ARM: an ARM64X forwarder per DLL sends ARM64 DAWs to the
  ARM64 DLL and x64 DAWs to the x64 one (one registry view for both).
  The Virtual Cable driver stays x64 only.
- Installer: all three architectures in one setup (HKLM32 entries,
  ARM64 files and forwarders only on ARM64, cable greyed out there).
- CI: windows-11-arm job builds and tests ARM64 and checks each
  forwarder from an ARM64 and an x64 process; the x64 job adds 32-bit.
- ADR 0016. Live on the x64 VM: 32-bit Master loop bit-exact, 32-bit
  Bridge client with a 64-bit Master carries audio.

+++ tests/network-test/main.cpp
@@ -138,8 +138,8 @@ RunResult RunLoopback(WHACodec codec, double seconds, uint32_t jitterMs, double 
   std::vector<float> outA(2 * kStride), inA(kStride), outB(kStride), inB(2 * kStride);
   const uint64_t ticks = static_cast<uint64_t>(seconds * kRate / kBlock);
   std::vector<float> got0, got1;
-  got0.reserve(ticks * kBlock);
-  got1.reserve(ticks * kBlock);
+  got0.reserve(static_cast<size_t>(ticks * kBlock));
+  got1.reserve(static_cast<size_t>(ticks * kBlock));
   LARGE_INTEGER freq, t0, now;
   QueryPerformanceFrequency(&freq);
   QueryPerformanceCounter(&t0);
